refactor(mips): simplify result handing on mips
MIPS/MIPS64 actually use the same technique as FreeBSD/macOS: they
return the result/error number directly and set a separate flag if an
error occurred. `sc` just hid this by re-encoding the error number on
error.

This means we can use the same code as FreeBSD/macOS to handle it, which
should result in speedups.